+++ src/main/java/ru/kata/spring/boot_security/demo/dao/UserDao.java
@@ -8,11 +8,11 @@ public interface UserDao {
 
     List<User> showAllUsers();
 
-    User showUserById(int id);
+    User showUserById(Long id);
 
     void saveUser(User user);
 
-    void updateUser(int id, User user);
+    void updateUser(Long id, User user);
 
-    void deleteUserById(int id);
+    void deleteUserById(Long id);
 }
